requirements.txt
fastapi==0.118.3
uvicorn==0.35.0
python-telegram-bot==22.5
google-genai==1.42.0
openai==2.3.0
python-dotenv==1.1.1
scikit-learn==1.7.2